QA: Flyttet endepunkt til egen controller
