+++ src/main/kotlin/no/nav/familie/tilbake/api/DokumentController.kt
@@ -15,12 +15,10 @@ import no.nav.familie.tilbake.dokumentbestilling.henleggelse.HenleggelsesbrevSer
 import no.nav.familie.tilbake.dokumentbestilling.varsel.VarselbrevService
 import no.nav.familie.tilbake.dokumentbestilling.vedtak.Avsnitt
 import no.nav.familie.tilbake.dokumentbestilling.vedtak.VedtaksbrevService
-import no.nav.familie.tilbake.faktaomfeilutbetaling.FaktaFeilutbetalingService
 import no.nav.familie.tilbake.sikkerhet.AuditLoggerEvent
 import no.nav.familie.tilbake.sikkerhet.Behandlerrolle
 import no.nav.familie.tilbake.sikkerhet.HenteParam
 import no.nav.familie.tilbake.sikkerhet.Rolletilgangssjekk
-import no.nav.familie.tilbake.vilkårsvurdering.VilkårsvurderingService
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.http.MediaType
 import org.springframework.web.bind.annotation.GetMapping
@@ -40,8 +38,6 @@ class DokumentController(
     private val henleggelsesbrevService: HenleggelsesbrevService,
     private val vedtaksbrevService: VedtaksbrevService,
     private val lagreUtkastVedtaksbrevService: LagreUtkastVedtaksbrevService,
-    private val faktaFeilutbetalingService: FaktaFeilutbetalingService,
-    private val vilkårsVurderingService: VilkårsvurderingService,
 ) {
     @Operation(summary = "Bestill brevsending")
     @PostMapping("/bestill")
@@ -132,23 +128,4 @@ class DokumentController(
         lagreUtkastVedtaksbrevService.lagreUtkast(behandlingId, fritekstavsnitt)
         return Ressurs.success("OK")
     }
-
-    @Operation(summary = "Sjekker om perioder er like - unntatt dato og beløp")
-    @PostMapping(
-        "/sjekk-likhet-perioder/{behandlingId}",
-        produces = [MediaType.APPLICATION_JSON_VALUE],
-    )
-    @Rolletilgangssjekk(
-        Behandlerrolle.SAKSBEHANDLER,
-        "Sjekker om perioder er like - unntatt dato og beløp",
-        AuditLoggerEvent.UPDATE,
-        HenteParam.BEHANDLING_ID,
-    )
-    fun erPerioderLike(
-        @PathVariable behandlingId: UUID,
-    ): Ressurs<Boolean> =
-        Ressurs.success(
-            faktaFeilutbetalingService.sjekkOmFaktaPerioderErLike(behandlingId) &&
-                vilkårsVurderingService.sjekkOmVilkårsvurderingPerioderErLike(behandlingId),
-        )
 }
